Added `ManualSynchronizationContextCore` type. Added `PromiseSynchronizationContext.Execute(bool exhaustive)` API.
