+++ Package/Core/Threading/ManualSynchronizationContextCore.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e3c41b0694d6d24e8b49aa35ec040ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
